optamize in to short int memory

+++ CPP/Engine/image_lib.cpp
@@ -5,15 +5,15 @@ class Image
     private:
         /* data */
     public:
-        int HEIGTH=0, WIDTH=0;
+        short int HEIGTH=0, WIDTH=0;
         Color** PIXCELS;
         
-        Image(int height, int width)
+        Image(short int height, short int width)
         {   this->HEIGTH = height; // Assigen Height and Width
             this->WIDTH = width;
             
             this->PIXCELS = new Color* [WIDTH]; // Create pixcels accorfing Height and Width
-            for (int i = 0; i < WIDTH; i++)
+            for (short int i = 0; i < WIDTH; i++)
             {
                 this->PIXCELS[i] = new Color [HEIGTH];
             }
